+++ libs/vre/resource-editor/resource-editor/src/lib/compound/compound-viewer.component.ts
@@ -1,11 +1,15 @@
 import { Component } from '@angular/core';
-import { Constants } from '@dasch-swiss/dsp-js';
+import { Constants, ReadStillImageFileValue } from '@dasch-swiss/dsp-js';
+import { filterUndefined } from '@dasch-swiss/vre/shared/app-common';
+import { map } from 'rxjs/operators';
 import { CompoundService } from './compound.service';
 
 @Component({
   selector: 'app-compound-viewer',
   template: `
     <ng-container *ngIf="compoundService.incomingResource$ | async as incomingResource">
+      <app-resource-legal *ngIf="fileValue$ | async as fileValue" [fileValue]="fileValue" />
+
       <app-still-image
         *ngIf="incomingResource.res.properties[HasStillImageFileValue]"
         class="dsp-representation"
@@ -16,6 +20,13 @@ import { CompoundService } from './compound.service';
 })
 export class CompoundViewerComponent {
   HasStillImageFileValue = Constants.HasStillImageFileValue;
+
+  fileValue$ = this.compoundService.incomingResource$.pipe(
+    filterUndefined(),
+    map(value => {
+      return value.res.properties[Constants.HasStillImageFileValue][0] as ReadStillImageFileValue;
+    })
+  );
 
   constructor(public compoundService: CompoundService) {}
 }