+++ libs/vre/shared/app-common-to-move/src/index.ts
@@ -3,7 +3,6 @@ export * from './lib/dialog-header.component';
 export * from './lib/filtered-resources.interface';
 export * from './lib/short-res-info.interface';
 export * from './lib/split.pipe';
-export * from './lib/incoming.service';
 export * from './lib/search-params.interface';
 export * from './lib/header/header.component';
 export * from './lib/status/status.component';